dynamodb-dump.py: Run under awscli v1 with boto3.
I had been running this script using the AWS python libraries included
with awscli (v2), but as of awscli-2.1.17 the activate_this.py module that
I was using has been moved and I did not get it to work from its new
location.  Instead I have changed the imports so that it will run under
the PyPI awscli (v1) and boto3 library.